+++ CHANGELOG.md
@@ -11,7 +11,18 @@
 
 ### ⭐ 新特性
 
-- **buession-logging-support-console：** 控制台日志处理器
+- **buession-logging-support-console：** 新增控制台日志处理
+
+
+### 🔔 变化
+
+- **buession-logging-support-jdbc：** 新增 LogDataConverter 实现数据和字段映射
+
+
+### 🐞 Bug 修复
+
+- **ALL：** 修复 ObjectProvider.ifUnique 的使用，应使用 ObjectProvider.ifAvailable
+
 
 ---